+++ client-course-schedulizer/src/components/pages/HarmonyPage/Harmony/Harmony.tsx
@@ -2,9 +2,6 @@ import { Harmony as HarmonyBase } from "@harmoniously/react";
 import React from "react";
 import { AssignmentsState, useAssignmentsStore } from "utilities";
 
-// TODO: allow this to be edited by users.
-// const times = ["mwf800", "mwf900", "mwf1030", "mwf1130", "mwf1230", "mwf130", "mwf230", "mwf330"];
-
 // TODO: move this to the harmony page and allow users to load in exiting data
 // temporary hook used to get data from the imported schedule and generate fake assignments.
 // const useInferredAssignments = () => {
@@ -32,8 +29,10 @@ const selector = ({ assignments }: AssignmentsState) => {
  *   gives metadata about the schedule.
  */
 export const Harmony = () => {
-  // const { inferredAssignments, professors, rooms } = useInferredAssignments();
   const assignments = useAssignmentsStore(selector);
+
+  // TODO: create a state to pass to this Component so I cen get the result back
+  // Then convert it to a CSV and upload it to the Schedulizer component.
 
   return (
     <>
