+++ client-course-schedulizer/src/components/pages/HarmonyPage/Harmony/HarmonyCourseChecboxes.tsx
@@ -1,0 +1,61 @@
+import { Box, Paper } from "@material-ui/core";
+import React, { useEffect, useState } from "react";
+import {
+  HarmonyAccessors,
+  HarmonyState,
+  SingularAccessors,
+  useAssignmentsStore,
+  useHarmonyStore,
+} from "utilities/hooks";
+import { HarmonyCheckboxList } from "./HarmonyCheckboxList";
+
+interface HarmonyCourseCheckboxesProps {
+  course: string;
+}
+
+// pick values from store.
+const selector = ({ professors, times, rooms }: HarmonyState) => {
+  return [professors, times, rooms];
+};
+
+/** All of the check box lists for each attribute for a specific class */
+export const HarmonyCourseCheckboxes = ({ course }: HarmonyCourseCheckboxesProps) => {
+  const [professors, times, rooms] = useHarmonyStore(selector);
+  const { setClass } = useAssignmentsStore();
+  const [profList, setProfList] = useState<string[]>([]);
+  const [timeList, setTimeList] = useState<string[]>([]);
+  const [roomList, setRoomList] = useState<string[]>([]);
+
+  useEffect(() => {
+    setClass(course, { professors: profList, rooms: roomList, times: timeList });
+  }, [course, profList, roomList, setClass, timeList]);
+
+  return (
+    <Box component="div" m={2}>
+      <Paper>
+        <b>{course}</b>
+        <HarmonyCheckboxList
+          course={course}
+          customLabel={(profObj: SingularAccessors["professor"]) => {
+            return `${profObj.First} ${profObj.Last}`;
+          }}
+          id="professors"
+          list={professors as HarmonyAccessors["professors"]}
+          setList={setProfList}
+        />
+        <HarmonyCheckboxList
+          course={course}
+          id="times"
+          list={times as HarmonyAccessors["times"]}
+          setList={setTimeList}
+        />
+        <HarmonyCheckboxList
+          course={course}
+          id="rooms"
+          list={rooms as HarmonyAccessors["rooms"]}
+          setList={setRoomList}
+        />
+      </Paper>
+    </Box>
+  );
+};
