+++ client-course-schedulizer/src/components/pages/HarmonyPage/Harmony/index.ts
@@ -1,2 +1,6 @@
 export * from "./Harmony";
+export * from "./HarmonyAddAssignments";
+export * from "./HarmonyCheckbox";
+export * from "./HarmonyCheckboxList";
+export * from "./HarmonyCourseChecboxes";
 export * from "./HarmonyFieldArrayForm";
